+++ src/lib/createFakeConnection.ts
@@ -1,7 +1,18 @@
-import { stub } from 'sinon';
 import * as typeorm from 'typeorm';
 import { FakeConnectionManager } from './FakeConnectionManager';
 import { DEFAULT_CONNECTION_NAME } from '@nestjs/typeorm/dist/typeorm.constants';
+
+export interface FakeConnectionOptions {
+  type: 'postgres';
+  name: string;
+  entities: (string | Function | typeorm.EntitySchema<any>)[]
+}
+
+export const defaultFakeConnectionOptions: FakeConnectionOptions = {
+  type: 'postgres',
+  name: DEFAULT_CONNECTION_NAME,
+  entities: []
+}
 
 // The fake connection, here we can get the queries and parameters.
 export const fakeConnection = {
@@ -18,18 +29,8 @@ export const getFakeConnectionManager = () => {
   return fakeConnectionManager;
 };
 
-export const createFakeConnection = stub(typeorm, 'createConnection').callsFake(
-  async function(optionsOrName: typeorm.ConnectionOptions) {
-    const connectionName =
-      typeof optionsOrName === 'string'
-        ? optionsOrName
-        : DEFAULT_CONNECTION_NAME;
-    const options =
-      optionsOrName instanceof Object
-        ? optionsOrName
-        : await typeorm.getConnectionOptions(connectionName);
-    return getFakeConnectionManager()
-      .create(options)
-      .connect();
-  },
-);
+export const createFakeConnection = (options: FakeConnectionOptions) => 
+  getFakeConnectionManager()
+    .create(options as typeorm.ConnectionOptions)
+    .connect();
+